requirement.txt
opencv-python
pillow
keras==2.1.3
tensorflow-gpu==1.14.0
numpy==1.16.6
imutils